Added minecraft advancement translations and new Meta Domain for Enahncement and Analysis.

+++ src/TheFipster.Minecraft.Import.Domain/PlayerStats.cs
@@ -6,25 +6,25 @@ namespace TheFipster.Minecraft.Import.Domain
     {
         public Stats()
         {
-            Killed = new Dictionary<string, double>();
-            KilledBy = new Dictionary<string, double>();
-            Used = new Dictionary<string, double>();
-            Crafted = new Dictionary<string, double>();
-            Custom = new Dictionary<string, double>();
-            Dropped = new Dictionary<string, double>();
-            Mined = new Dictionary<string, double>();
-            PickedUp = new Dictionary<string, double>();
-            Broken = new Dictionary<string, double>();
+            Killed = new Dictionary<string, int>();
+            KilledBy = new Dictionary<string, int>();
+            Used = new Dictionary<string, int>();
+            Crafted = new Dictionary<string, int>();
+            Custom = new Dictionary<string, int>();
+            Dropped = new Dictionary<string, int>();
+            Mined = new Dictionary<string, int>();
+            PickedUp = new Dictionary<string, int>();
+            Broken = new Dictionary<string, int>();
         }
 
-        public Dictionary<string, double> Killed { get; set; }
-        public Dictionary<string, double> KilledBy { get; set; }
-        public Dictionary<string, double> Used { get; set; }
-        public Dictionary<string, double> Crafted { get; set; }
-        public Dictionary<string, double> Custom { get; set; }
-        public Dictionary<string, double> Dropped { get; set; }
-        public Dictionary<string, double> Mined { get; set; }
-        public Dictionary<string, double> PickedUp { get; set; }
-        public Dictionary<string, double> Broken { get; set; }
+        public Dictionary<string, int> Killed { get; set; }
+        public Dictionary<string, int> KilledBy { get; set; }
+        public Dictionary<string, int> Used { get; set; }
+        public Dictionary<string, int> Crafted { get; set; }
+        public Dictionary<string, int> Custom { get; set; }
+        public Dictionary<string, int> Dropped { get; set; }
+        public Dictionary<string, int> Mined { get; set; }
+        public Dictionary<string, int> PickedUp { get; set; }
+        public Dictionary<string, int> Broken { get; set; }
     }
 }
